fix incorrect use of <expr> mappings
Defining a mapping with 'map <expr> {lhs} {rhs}' means when {lhs} is typed,
{rhs} will be evaluated. Use this when you want to map {lhs} to the return value
of the evaluated expression.

I was incorrectly using <expr> to call functions without causing ':call ...' to
show up on the command line. This happens to work in some cases, but not all
(calling ReloadBuffers this way was not working).

When the mapping is concerned with evaluating an expression or getting a
function'n return value, use <expr>. When it's concerned with triggering a
function for its side-effects, use <silent> and ':call'