add Super-Shell-Indent plugin
This seems to get shell indentation right where the stock indentation doesn't. For
example, it yields

  some_array=(
    foo
    bar
  )

where the stock indentation yields

  some_array=(
  foo
  bar
  )

The readme mentions it's slower than the stock indentation, so I'll keep an eye on
performance and see if it's actually noticeable. I mean, it's not Emacs...

For future reference, this plugin is very very very old. I see there's a more recent
contender at
  https://github.com/vim-scripts/sh.vim--Cla
though the code doesn't look quite as clean at a glance. Worth keeping in mind though if
this one exhibits problems.