sh.vim syntax: allow { expression-list; } to fold
I copied this line from the existing syntax/sh.vim file in the runtime
dir, and simply added the `fold` argument to allow folding code like,

  { # <- fold here
    echo 'welcome to the jungle'

    foo() {
      ...
    }
  }

If this was to be proposed in a PR to the original, I'd probably need to
find a way to work it into that file's existing framework of
user-options (e.g. the g:sh_fold_enabled bitmask).