keybindings: learn some lessons from vim-rsi
See https://github.com/tpope/vim-rsi
The idea of making a keybinding conditional on whether the cursor is at the
end of the line is clever, but problematic in cases where key-repeat
could move you to the end of the line, triggering unwanted behavior if
you don't release the key in time. How critical this is depends on the
consequences of the unintented behavior.

For example, in insert mode, we could use <C-D> for forward delete when within
the line and revert to the original de-indent behavior when the cursor's at the
end of the line. But if we hold down <C-D> within the line to delete a
bunch of following characters, the cursor could reach the end of the
line, at which point we'd unintentionally trigger de-indenting.

In command-line mode, on the other hand, holding down <C-D> could bring
the cursor to the end of the line, but the worst that happens if you hit
<C-D> again is you'll see completion candidates displayed.

For this reason, I chose to apply conditional behavior to some mappings (e.g.
c_CTRL-D) and not to others (e.g. i_CTRL-D, c_CTRL-F).