tags, articulate.vim, iskeyword
Provide a solution to the vimscript iskeyword-=# conundrum, via a
wrapper that temporarily sets iskeyword+=# while performing a tag
lookup. Decouple the keybindings I'm setting in vimrc from the bindings
I need to reference in the ftplugin file by extracting an articulate.vim
plugin that defines a stable set of <Plug> bindings.