Space leader: avoid <Nop> timeout bug
I use <Space> as my leader key, but don't want the cursor to move
(default Space behavior) if I let the timeout expire partway through
typing a leader sequence. One way to prevent this is to map Space to
<Nop>, but then if the timeout expires, for some reason Vim ceases to
recognize a newly started leader sequence (it's like it's still running
the Nop) until you run some intervening command (e.g. type `l`, or `h`,
or `Esc`).

Instead, leave mapleader at it's default `\` and map <Space> to
<Leader>. This has the benefit that `\` shows up in the showcmd area
when entering a leader sequence. The downside is we're occupying extra
keyboard real-estate, but in practice I never use `\` directly anyway,
as it's an awkward key to reach for.