improve <leader><C-H> help binding
Putting the cursor at the ent of the line is more useful: append a ' to specify
an option, append ( for a function, hit Tab to see completions. The only
use-case this is worse for is prepending a : before the current word to specify
an ex command.

Also, there's no need to manually yank and paste the current word, use
<C-R><C-W> for that.