rubyfuncs#align_chain()
Write a function that aligns a chained Ruby method call to the previous
link in the chain, transforming something like,

  foobaristical.jay
    .walk

into

  foobaristical.jay
               .walk

This only aligns the line the cursor is currently on. A nice next step
would be to make it align the full range of lines that the method chain
spans.