vimruby#align_method_chain()
Provide a function to align a multi-line ruby method chain, turning
this:

  feefifo.foo
    .goo
    .gum

into this:

  feefifo.foo
         .goo
         .gum

In its current form, this only aligns the current line and any previous
lines in the chain. A good next step would be to apply alignment down
to subsequent lines as well, so you could trigger the function anywhere
within the chain.