Ruby: handle trailing ! and ? in ctag lookups
I seem to have stumbled upon an undocumented gem here. Because ! and ?
are not among the 'iskeyword' characters, attempting a tag lookup by
hitting <C-]> with the cursor on such a method name fails, because Vim
looks up the name without the punctuation.

Adding these characters to the 'iskeyword' option causes other problems,
such as inclusion of leading ! negation in tag lookups. As you'd expect,
evaluating `expand('<cword>')` omits the punctuation if it's not a
keyword character.

However, for some reason, and I can't find any reference to this in the
documentation, typing <C-R><C-W> from command-line mode expands the word
under the cursor including the trailing punctuation, even if that's not
a keyword character. This provides a perfect, simple solution to Ruby
tag lookups.