tags: remove some imperfect mapping examples
Dealing with visual mode requires a more nuanced approach, including clearing
the command line with <C-U> and grabbing the selected range to perform the tag
lookup. But if you're manually selecting the characters to lookup, there's no
need to jump through any <C-R><C-W> hoops to infer which characters to look up!