go native (Vim 8 package feature)
Remove vim-plug and use Vim 8's native packages feature. I've moved most
of the plugins into my nixos config, but there are a few I'm not ready
to do that with. These, I've moved into
~/.vim/after/pack/miscpack/{opt,start}

A downside of this move is that those packages are no longer reflected
in the version-control, but this is a temporary solution anyway -- I
want to spend some time addressing these plugins individually to sort
out the issues I have with them and move stuff into my nixos config
where it makes sense.