move after/{ftplugin,plugin,syntax} into ftglue
I was wrong in thinking that Vim doesn't load files from after/
directories in plugins -- it will load files from any after/ directories
in runtimepath at the appropriate time, but aside from loading at a
later time, it treats after/ directories like any other directories in
the runtimepath, meaning it expects the same directory structure. So we
can make use of an after/ directory in a plugin as long as we structure
that directory accordingly. I feel like I'm talking in circles, but I
think the contents of the following commit demonstrate what I'm trying
to say:

https://github.com/ivanbrennan/ftglue/commit/b346b35eaef7b6010fe3fcaafd69383599348fe1