setup haskell-vim syntax before filetype.vim loads
This needs to be setup before $VIMRUNTIME/filetype.vim is sourced, since
that file defines an autocommand in the same augroup that would load
$VIMRUNTIME/syntax/haskell.vim instead. Adding an earlier autocommand
allows us to load the plugin's syntax file first, even though it occurs
later in the runtimepath.

An interesting note is that attempting to put this initialization in any
other file (e.g. vimrc.local) won't work, because the moment we try to
source that file, Vim will first source filetype.vim.